separate PL1 and PL2
