+++ opensr_srgan/model/SRGAN.py
@@ -54,8 +54,13 @@ class SRGAN_model(pl.LightningModule):
             raise TypeError("Config must be a filepath (str or Path), dict, or OmegaConf object.")
         assert mode in {"train", "eval"}, "Mode must be 'train' or 'eval'"  # validate mode
         
+        
+        # Set attributes
         self.config = config
-        self.mode = mode                                # store mode (train/eval)
+        self.mode = mode          
+        
+        # Get PL version
+        self.pl_version = tuple(int(x) for x in pl.__version__.split("."))
 
         # --- Training settings ---
         self.pretrain_g_only = bool(getattr(self.config.Training, "pretrain_g_only", False))  # pretrain generator only (default False)
@@ -227,91 +232,15 @@ class SRGAN_model(pl.LightningModule):
         # Purpose: compute SR prediction, evaluate training metrics, log them.
         # ======================================================================
 
-        # -------- CREATE SR DATA --------
-        lr_imgs, hr_imgs = batch                                  # unpack LR/HR tensors from dataloader batch
-        sr_imgs = self.forward(lr_imgs)                          # forward pass of the generator to produce SR from LR
-
-        # ======================================================================
-        # SECTION: Pretraining phase gate
-        # Purpose: decide if we are in the content-only pretrain stage.
-        # ======================================================================
-
-        # -------- DETERMINE PRETRAINING --------
-        pretrain_phase = self._pretrain_check()                  # check schedule: True => content-only pretraining
-        if optimizer_idx == 1:  # log whether pretraining is active or not
-            self.log("training/pretrain_phase", float(pretrain_phase), prog_bar=False,sync_dist=True)  # log once per G step to track phase state
-
-        # ======================================================================
-        # SECTION: Pretraining branch (delegated)
-        # Purpose: during pretrain, only content loss for G and dummy logging for D.
-        # ======================================================================
-
-        # -------- IF PRETRAIN: delegate --------
-        if pretrain_phase:
-            # run pretrain step separately and return loss here
-            return self.pretraining_training_step(lr_imgs=lr_imgs, hr_imgs=hr_imgs, sr_imgs=sr_imgs, optimizer_idx=optimizer_idx)  # delegate pretrain logic (no forward here)
-
-        # ======================================================================
-        # SECTION: Adversarial training — Discriminator step
-        # Purpose: update D to distinguish HR (real) vs SR (fake).
-        # ======================================================================
-
-        # -------- Normal Train: Discriminator Step  --------
-        if optimizer_idx==0:
-            # run discriminator and get loss between pred labels and true labels
-            hr_discriminated = self.discriminator(hr_imgs)       # D(real): logits for HR images
-            sr_discriminated = self.discriminator(sr_imgs.detach()) # detach so G doesn’t get gradients from D’s step
-
-            # targets
-            real_target = torch.full_like(hr_discriminated, self.adv_target) # get labels/fuzzy labels
-            fake_target = torch.zeros_like(sr_discriminated) # zeros, since generative prediction
-
-            # Binary Cross-Entropy loss
-            loss_real = self.adversarial_loss_criterion(hr_discriminated, real_target)   # BCEWithLogitsLoss for D(G(x))
-            loss_fake = self.adversarial_loss_criterion(sr_discriminated, fake_target)  # BCEWithLogitsLoss for D(y)
-            adversarial_loss = loss_real + loss_fake # Sum up losses
-            self.log("discriminator/adversarial_loss",adversarial_loss,sync_dist=True) # log weighted loss
-
-            # [LOG-B] Always log D opinions: real probs in normal training
-            with torch.no_grad():
-                d_real_prob = torch.sigmoid(hr_discriminated).mean()   # estimate mean real probability
-                d_fake_prob = torch.sigmoid(sr_discriminated).mean()   # estimate mean fake probability
-            self.log("discriminator/D(y)_prob", d_real_prob, prog_bar=True,sync_dist=True)      # log D(real) confidence
-            self.log("discriminator/D(G(x))_prob", d_fake_prob, prog_bar=True,sync_dist=True)   # log D(fake) confidence
-
-            # return weighted discriminator loss
-            return adversarial_loss                                # PL will use this to step the D optimizer
-
-        # ======================================================================
-        # SECTION: Adversarial training — Generator step
-        # Purpose: update G to minimize content loss + (weighted) adversarial loss.
-        # ======================================================================
-
-        # -------- Normal Train: Generator Step  --------
-        if optimizer_idx==1:
-
-            """ 1. Get VGG space loss """
-            # encode images
-            content_loss, metrics = self.content_loss_criterion.return_loss(sr_imgs, hr_imgs)   # perceptual/content criterion (e.g., VGG)
-            self._log_generator_content_loss(content_loss)                             # log content loss for G (consistent args)
-            for key, value in metrics.items():
-                self.log(f"train_metrics/{key}", value,sync_dist=True)                             # log detailed metrics without extra forward passes
-
-
-            """ 2. Get Discriminator Opinion and loss """
-            # run discriminator and get loss between pred labels and true labels
-            sr_discriminated = self.discriminator(sr_imgs)                             # D(SR): logits for generator outputs
-            adversarial_loss = self.adversarial_loss_criterion(sr_discriminated, torch.ones_like(sr_discriminated)) # keep taargets 1.0 for G loss
-            self.log("generator/adversarial_loss",adversarial_loss,sync_dist=True)     # log unweighted adversarial loss
-
-            """ 3. Weight the losses"""
-            adv_weight = self._adv_loss_weight() # get adversarial weight based on current step
-            adversarial_loss_weighted = (adversarial_loss * adv_weight) # weight adversarial loss
-            total_loss = content_loss + adversarial_loss_weighted # total content loss
-            self.log("generator/total_loss",total_loss,sync_dist=True)                # log combined objective (content + λ_adv * adv)
-
-            # return Generator loss
-            return total_loss                                                         # PL will use this to step the G optimizer
+        # Check for PL version - define training step accordingly
+        if self.pl_version >= (2,0,0):
+            from opensr_srgan.model.training_step_PL import training_step_PL2x
+            return training_step_PL2x(self, batch, batch_idx, optimizer_idx)
+        elif self.pl_version < (2,0,0):
+            from opensr_srgan.model.training_step_PL import training_step_PL1x
+            return training_step_PL1x(self, batch, batch_idx, optimizer_idx)
+        
+        
 
     def optimizer_step(
         self,
@@ -333,46 +262,6 @@ class SRGAN_model(pl.LightningModule):
             and self.global_step >= self._ema_update_after_step
         ):
             self.ema.update(self.generator)
-
-    def pretraining_training_step(self, *, lr_imgs, hr_imgs, sr_imgs, optimizer_idx):
-        """
-        Pretraining branch logic (no forward here):
-        - G step (optimizer_idx==1): content-only loss
-        - D step (optimizer_idx==0): dummy loss + zeroed logs so closures run
-        """
-
-        # ======================================================================
-        # SECTION: Generator (G) pretraining step
-        # Purpose: train only the generator using content loss, no adversarial part.
-        # ======================================================================
-        if optimizer_idx == 1:
-            content_loss, metrics = self.content_loss_criterion.return_loss(sr_imgs, hr_imgs)  # compute perceptual/content loss (e.g., VGG or L1)
-            self._log_generator_content_loss(content_loss)                             # log content loss for G (consistent args)
-            for key, value in metrics.items():
-                self.log(f"train_metrics/{key}", value,sync_dist=True)                               # reuse computed metrics for logging
-
-            # Ensure adversarial weight is logged even when not used during pretraining
-            adv_weight = self._compute_adv_loss_weight()
-            self._log_adv_loss_weight(adv_weight)
-            return content_loss                                                        # return loss for optimizer step (G only)
-
-        # ======================================================================
-        # SECTION: Discriminator (D) pretraining step
-        # Purpose: no real training — just log zeros and return dummy loss to satisfy closure.
-        # ======================================================================
-        elif optimizer_idx == 0:
-            device, dtype = hr_imgs.device, hr_imgs.dtype                              # get tensor device and dtype for consistency
-            zero = torch.tensor(0.0, device=device, dtype=dtype)                       # define reusable zero tensor
-
-            # --- Log dummy discriminator "opinions" (always zero during pretrain) ---
-            self.log("discriminator/D(y)_prob",    zero, prog_bar=True,  sync_dist=True)  # fake real-prob (always 0)
-            self.log("discriminator/D(G(x))_prob", zero, prog_bar=True,  sync_dist=True)  # fake fake-prob (always 0)
-
-            # --- Create dummy scalar loss (ensures PL closure runs) ---
-            dummy = torch.zeros((), device=device, dtype=dtype, requires_grad=True)    # dummy value with grad for optimizer compatibility
-            self.log("discriminator/adversarial_loss", dummy, sync_dist=True)          # log dummy adversarial loss (always 0)
-            return dummy                                                               # return dummy loss (keeps Lightning loop intact)
-
 
     @torch.no_grad()
     def validation_step(self, batch, batch_idx):
@@ -472,76 +361,14 @@ class SRGAN_model(pl.LightningModule):
 
     def configure_optimizers(self):
 
-        # configure Generator optimizer (Adam)
-        optimizer_g = torch.optim.Adam(
-            params=filter(lambda p: p.requires_grad, self.generator.parameters()),  # only trainable params
-            lr=self.config.Optimizers.optim_g_lr                                     # LR from config
-        )
-
-        # configure Discriminator optimizer (Adam)
-        optimizer_d = torch.optim.Adam(
-            params=filter(lambda p: p.requires_grad, self.discriminator.parameters()),  # only trainable params
-            lr=self.config.Optimizers.optim_d_lr                                       # LR from config
-        )
-
-        # learning rate schedulers (ReduceLROnPlateau)
-        scheduler_g = ReduceLROnPlateau(
-            optimizer_g, mode='min',
-            factor=self.config.Schedulers.factor_g,
-            patience=self.config.Schedulers.patience_g,
-            #verbose=self.config.Schedulers.verbose
-        )
-        scheduler_d = ReduceLROnPlateau(
-            optimizer_d, mode='min',
-            factor=self.config.Schedulers.factor_d,
-            patience=self.config.Schedulers.patience_d,
-            #verbose=self.config.Schedulers.verbose
-        )
-
-        # optional generator warmup scheduler (step-based)
-        warmup_steps = int(getattr(self.config.Schedulers, "g_warmup_steps", 0))
-        warmup_type = getattr(self.config.Schedulers, "g_warmup_type", "none").lower()
-
-        warmup_scheduler_config = None
-        if warmup_steps > 0 and warmup_type in {"cosine", "linear"}:
-
-            def _generator_warmup_lambda(current_step: int) -> float:
-                if current_step >= warmup_steps:
-                    return 1.0
-
-                progress = (current_step + 1) / float(max(1, warmup_steps))
-                if warmup_type == "linear":
-                    return progress
-
-                # default to cosine warmup for smoother start
-                return 0.5 * (1.0 - math.cos(math.pi * progress))
-
-            warmup_scheduler = torch.optim.lr_scheduler.LambdaLR(
-                optimizer_g,
-                lr_lambda=_generator_warmup_lambda,
-            )
-
-            warmup_scheduler_config = {
-                'scheduler': warmup_scheduler,
-                'interval': 'step',
-                'frequency': 1,
-                'name': 'generator_warmup',
-            }
-
-        scheduler_configs = [
-            {'scheduler': scheduler_d, 'monitor': self.config.Schedulers.metric, 'reduce_on_plateau': True, 'interval': 'epoch', 'frequency': 1},
-            {'scheduler': scheduler_g, 'monitor': self.config.Schedulers.metric, 'reduce_on_plateau': True, 'interval': 'epoch', 'frequency': 1}
-        ]
-
-        if warmup_scheduler_config is not None:
-            scheduler_configs.append(warmup_scheduler_config)
-
-        # return both optimizers + schedulers for PL
-        return [
-            [optimizer_d, optimizer_g],  # order super important, it's [D, G] and checked in training step
-            scheduler_configs,
-        ]
-
+        # check for PL version - define optimisers accordingly
+        if self.pl_version >= (2,0,0):
+            from opensr_srgan.model.configure_optimizers_PL import configure_optimizers_PL2x
+            return configure_optimizers_PL2x(self)
+        elif self.pl_version <= (2,0,0):
+            from opensr_srgan.model.configure_optimizers_PL import configure_optimizers_PL1x
+            return configure_optimizers_PL1x(self)
+            
 
     def on_train_batch_start(self, batch, batch_idx):  # called before each training batch
         pre = self._pretrain_check()                   # check if currently in pretraining phase
